Add limited sight to the player
In this commit we are introducing the sight of the player.

The sight is a radius around the player that changes the visibility of
the objects, if an object is inside of this radius it is displayed, and
it is hidden otherwise.

In order to add some effect, the visibility of the objects is
proportional to its distance from the player.

All the elements are now created by a function called createElement,
which makes the code cleaner and easier to change.